# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Your API Title","description":"This is a sample Petstore server. You can find more about Swagger at [http:\/\/swagger.io](http:\/\/swagger.io) or on [irc.freenode.net, #swagger](http:\/\/swagger.io\/irc\/).","termsOfService":"http:\/\/swagger.io\/terms\/","contact":{"email":"apiteam@swagger.io"},"license":{"name":"Apache 2.0","url":"http:\/\/www.apache.org\/licenses\/LICENSE-2.0.html"},"version":"1.0.0"},"paths":{"\/login":{"post":{"tags":["X\u00e1c th\u1ef1c"],"summary":"\u0110\u0103ng nh\u1eadp ng\u01b0\u1eddi d\u00f9ng","operationId":"2dd578ff7aba2721293dac66833b27d1","requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["email","password"],"properties":{"email":{"type":"string","example":"janesmith@example.com"},"password":{"type":"string","example":"securepassword"}},"type":"object"}}}},"responses":{"200":{"description":"\u0110\u0103ng nh\u1eadp th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"\u0110\u0103ng nh\u1eadp th\u00e0nh c\u00f4ng!"},"data":{"properties":{"access_token":{"type":"string","example":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."},"token_type":{"type":"string","example":"bearer"},"role":{"type":"integer","example":1},"expires_in":{"type":"integer","example":3600}},"type":"object"}},"type":"object"}}}},"401":{"description":"Th\u00f4ng tin \u0111\u0103ng nh\u1eadp kh\u00f4ng h\u1ee3p l\u1ec7","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i \u0111\u0103ng nh\u1eadp"},"error":{"type":"string","example":"Email ho\u1eb7c m\u1eadt kh\u1ea9u kh\u00f4ng ch\u00ednh x\u00e1c"}},"type":"object"}}}},"422":{"description":"L\u1ed7i x\u00e1c th\u1ef1c d\u1eef li\u1ec7u","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i x\u00e1c th\u1ef1c"},"errors":{"type":"object"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}}}},"\/logout":{"post":{"tags":["X\u00e1c th\u1ef1c"],"summary":"\u0110\u0103ng xu\u1ea5t ng\u01b0\u1eddi d\u00f9ng","operationId":"68a6108af04af5e3bc1ea68a6c6f1299","responses":{"200":{"description":"\u0110\u0103ng xu\u1ea5t th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"\u0110\u0103ng xu\u1ea5t th\u00e0nh c\u00f4ng"}},"type":"object"}}}},"401":{"description":"Ch\u01b0a x\u00e1c th\u1ef1c","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"Ch\u01b0a x\u00e1c th\u1ef1c"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"Kh\u00f4ng th\u1ec3 \u0111\u0103ng xu\u1ea5t"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"\/admin\/customers":{"get":{"tags":["Qu\u1ea3n l\u00fd kh\u00e1ch h\u00e0ng"],"summary":"L\u1ea5y danh s\u00e1ch kh\u00e1ch h\u00e0ng","operationId":"e85f001e292b07d069462e2ea6d4ac35","parameters":[{"name":"search","in":"query","description":"T\u00ecm ki\u1ebfm theo t\u00ean ho\u1eb7c email kh\u00e1ch h\u00e0ng","required":false,"schema":{"type":"string","example":"Nguyen Van A"}},{"name":"sort","in":"query","description":"S\u1eafp x\u1ebfp theo ng\u00e0y t\u1ea1o (asc ho\u1eb7c desc)","required":false,"schema":{"type":"string","example":"desc"}},{"name":"per_page","in":"query","description":"S\u1ed1 l\u01b0\u1ee3ng kh\u00e1ch h\u00e0ng tr\u00ean m\u1ed7i trang","required":false,"schema":{"type":"integer","example":10}}],"responses":{"200":{"description":"L\u1ea5y danh s\u00e1ch kh\u00e1ch h\u00e0ng th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"L\u1ea5y danh s\u00e1ch th\u00e0nh c\u00f4ng"},"data":{"properties":{"current_page":{"type":"integer","example":1},"data":{"type":"array","items":{"properties":{"id":{"type":"integer","example":1},"name":{"type":"string","example":"Nguyen Van A"},"email":{"type":"string","example":"example@gmail.com"},"created_at":{"type":"string","example":"2025-03-31 12:00:00"},"updated_at":{"type":"string","example":"2025-03-31 12:00:00"}},"type":"object"}},"total":{"type":"integer","example":100},"per_page":{"type":"integer","example":10},"last_page":{"type":"integer","example":10}},"type":"object"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Qu\u1ea3n l\u00fd kh\u00e1ch h\u00e0ng"],"summary":"T\u1ea1o kh\u00e1ch h\u00e0ng m\u1edbi","operationId":"7ef2da549c406a9cab74f0dab03c2d8a","requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["name","email","password"],"properties":{"name":{"type":"string","example":"Nguy\u1ec5n V\u0103n A"},"email":{"type":"string","example":"example@gmail.com"},"password":{"type":"string","example":"password123"}},"type":"object"}}}},"responses":{"201":{"description":"T\u1ea1o kh\u00e1ch h\u00e0ng th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"T\u1ea1o t\u00e0i kho\u1ea3n th\u00e0nh c\u00f4ng!"},"data":{"type":"object"}},"type":"object"}}}},"422":{"description":"D\u1eef li\u1ec7u kh\u00f4ng h\u1ee3p l\u1ec7 - L\u1ed7i x\u00e1c th\u1ef1c","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i x\u00e1c th\u1ef1c"},"errors":{"type":"object"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"\/admin\/orders":{"get":{"tags":["Qu\u1ea3n l\u00fd kh\u00e1ch h\u00e0ng"],"summary":"L\u1ea5y t\u1ea5t c\u1ea3 \u0111\u01a1n h\u00e0ng v\u00e0 th\u00f4ng tin kh\u00e1ch h\u00e0ng","operationId":"b5d14fe0c552a402100411a3cc042961","parameters":[{"name":"order_id","in":"query","description":"ID c\u1ee7a \u0111\u01a1n h\u00e0ng \u0111\u1ec3 t\u00ecm ki\u1ebfm","required":false,"schema":{"type":"integer","example":1}},{"name":"name","in":"query","description":"T\u00ean kh\u00e1ch h\u00e0ng \u0111\u1ec3 t\u00ecm ki\u1ebfm (kh\u1edbp m\u1ed9t ph\u1ea7n)","required":false,"schema":{"type":"string","example":"Nguyen"}},{"name":"email","in":"query","description":"Email kh\u00e1ch h\u00e0ng \u0111\u1ec3 t\u00ecm ki\u1ebfm (kh\u1edbp m\u1ed9t ph\u1ea7n)","required":false,"schema":{"type":"string","example":"example@gmail.com"}},{"name":"per_page","in":"query","description":"S\u1ed1 l\u01b0\u1ee3ng \u0111\u01a1n h\u00e0ng tr\u00ean m\u1ed7i trang","required":false,"schema":{"type":"integer","example":10}}],"responses":{"200":{"description":"L\u1ea5y danh s\u00e1ch \u0111\u01a1n h\u00e0ng th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"L\u1ea5y danh s\u00e1ch \u0111\u01a1n h\u00e0ng th\u00e0nh c\u00f4ng"},"data":{"properties":{"current_page":{"type":"integer","example":1},"data":{"type":"array","items":{"properties":{"id":{"type":"integer","example":1},"total":{"type":"number","example":150000},"created_at":{"type":"string","example":"2025-03-31 12:00:00"},"customer":{"properties":{"id":{"type":"integer","example":1},"name":{"type":"string","example":"Nguyen Van A"},"email":{"type":"string","example":"example@gmail.com"}},"type":"object"}},"type":"object"}},"total":{"type":"integer","example":50},"per_page":{"type":"integer","example":10},"last_page":{"type":"integer","example":5}},"type":"object"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"\/admin\/customers\/{id}":{"put":{"tags":["Qu\u1ea3n l\u00fd kh\u00e1ch h\u00e0ng"],"summary":"C\u1eadp nh\u1eadt th\u00f4ng tin kh\u00e1ch h\u00e0ng","operationId":"a3dd9c018cd08d38a44ddb33b2ecdbfd","parameters":[{"name":"id","in":"path","description":"ID c\u1ee7a kh\u00e1ch h\u00e0ng","required":true,"schema":{"type":"integer","example":1}}],"requestBody":{"required":true,"content":{"application\/json":{"schema":{"properties":{"name":{"type":"string","example":"Nguy\u1ec5n V\u0103n A"}},"type":"object"}}}},"responses":{"201":{"description":"C\u1eadp nh\u1eadt kh\u00e1ch h\u00e0ng th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"C\u1eadp nh\u1eadt th\u00e0nh c\u00f4ng"},"data":{"type":"object"}},"type":"object"}}}},"404":{"description":"Kh\u00f4ng t\u00ecm th\u1ea5y kh\u00e1ch h\u00e0ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"Kh\u00f4ng t\u00ecm th\u1ea5y kh\u00e1ch h\u00e0ng"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Qu\u1ea3n l\u00fd kh\u00e1ch h\u00e0ng"],"summary":"X\u00f3a kh\u00e1ch h\u00e0ng","operationId":"9efa82774bd5fd7a1ecf43dbf0f317d0","parameters":[{"name":"id","in":"path","description":"ID c\u1ee7a kh\u00e1ch h\u00e0ng","required":true,"schema":{"type":"integer","example":1}}],"responses":{"200":{"description":"X\u00f3a kh\u00e1ch h\u00e0ng th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"X\u00f3a kh\u00e1ch h\u00e0ng th\u00e0nh c\u00f4ng"}},"type":"object"}}}},"404":{"description":"Kh\u00f4ng t\u00ecm th\u1ea5y kh\u00e1ch h\u00e0ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"Kh\u00f4ng t\u00ecm th\u1ea5y kh\u00e1ch h\u00e0ng"}},"type":"object"}}}},"400":{"description":"Kh\u00f4ng th\u1ec3 x\u00f3a kh\u00e1ch h\u00e0ng c\u00f3 \u0111\u01a1n h\u00e0ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"Kh\u00f4ng th\u1ec3 x\u00f3a kh\u00e1ch h\u00e0ng c\u00f3 \u0111\u01a1n h\u00e0ng"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"\/admin\/change-password":{"put":{"tags":["Qu\u1ea3n l\u00fd kh\u00e1ch h\u00e0ng"],"summary":"Thay \u0111\u1ed5i m\u1eadt kh\u1ea9u c\u1ee7a admin","operationId":"dd2545f1960cf248177aeeaa5392a3cd","requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["old_password","new_password"],"properties":{"old_password":{"type":"string","example":"123456789"},"new_password":{"type":"string","example":"newpassword123"}},"type":"object"}}}},"responses":{"201":{"description":"Thay \u0111\u1ed5i m\u1eadt kh\u1ea9u th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"Thay \u0111\u1ed5i m\u1eadt kh\u1ea9u th\u00e0nh c\u00f4ng"}},"type":"object"}}}},"401":{"description":"Ch\u01b0a x\u00e1c th\u1ef1c ho\u1eb7c m\u1eadt kh\u1ea9u c\u0169 kh\u00f4ng \u0111\u00fang","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"M\u1eadt kh\u1ea9u c\u0169 kh\u00f4ng \u0111\u00fang"}},"type":"object"}}}},"422":{"description":"L\u1ed7i x\u00e1c th\u1ef1c d\u1eef li\u1ec7u","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i x\u00e1c th\u1ef1c"},"errors":{"type":"object"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"\/customer\/info":{"get":{"tags":["Th\u00f4ng tin kh\u00e1ch h\u00e0ng"],"summary":"L\u1ea5y th\u00f4ng tin kh\u00e1ch h\u00e0ng hi\u1ec7n t\u1ea1i","operationId":"b14c01a98dabdb1864aeff83514bfb01","responses":{"201":{"description":"L\u1ea5y th\u00f4ng tin kh\u00e1ch h\u00e0ng th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"L\u1ea5y th\u00f4ng tin th\u00e0nh c\u00f4ng"},"data":{"type":"object"}},"type":"object"}}}},"401":{"description":"Ch\u01b0a x\u00e1c th\u1ef1c","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"Ch\u01b0a x\u00e1c th\u1ef1c"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"\/customer":{"put":{"tags":["Th\u00f4ng tin kh\u00e1ch h\u00e0ng"],"summary":"C\u1eadp nh\u1eadt th\u00f4ng tin c\u00e1 nh\u00e2n kh\u00e1ch h\u00e0ng hi\u1ec7n t\u1ea1i","operationId":"26cb2da7ca6856e63c7fc1611c20041f","requestBody":{"required":true,"content":{"application\/json":{"schema":{"properties":{"name":{"type":"string","example":"Nguy\u1ec5n V\u0103n A"}},"type":"object"}}}},"responses":{"201":{"description":"C\u1eadp nh\u1eadt th\u00f4ng tin th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"C\u1eadp nh\u1eadt th\u00f4ng tin th\u00e0nh c\u00f4ng"},"data":{"type":"object"}},"type":"object"}}}},"401":{"description":"Ch\u01b0a x\u00e1c th\u1ef1c","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"Ch\u01b0a x\u00e1c th\u1ef1c"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"\/customer\/change-password":{"put":{"tags":["Th\u00f4ng tin kh\u00e1ch h\u00e0ng"],"summary":"Thay \u0111\u1ed5i m\u1eadt kh\u1ea9u c\u1ee7a kh\u00e1ch h\u00e0ng","operationId":"74d5e39161f73efd5a82e5c114cb5f5a","requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["old_password","new_password"],"properties":{"old_password":{"type":"string","example":"password123"},"new_password":{"type":"string","example":"newpassword456"}},"type":"object"}}}},"responses":{"201":{"description":"Thay \u0111\u1ed5i m\u1eadt kh\u1ea9u th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"Thay \u0111\u1ed5i m\u1eadt kh\u1ea9u th\u00e0nh c\u00f4ng"}},"type":"object"}}}},"401":{"description":"Ch\u01b0a x\u00e1c th\u1ef1c ho\u1eb7c m\u1eadt kh\u1ea9u c\u0169 kh\u00f4ng \u0111\u00fang","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"M\u1eadt kh\u1ea9u c\u0169 kh\u00f4ng \u0111\u00fang"}},"type":"object"}}}},"422":{"description":"L\u1ed7i x\u00e1c th\u1ef1c d\u1eef li\u1ec7u","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i x\u00e1c th\u1ef1c"},"errors":{"type":"object"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"\/customer\/orders":{"get":{"tags":["Danh s\u00e1ch \u0111\u01a1n h\u00e0ng"],"summary":"L\u1ea5y danh s\u00e1ch \u0111\u01a1n h\u00e0ng c\u1ee7a kh\u00e1ch h\u00e0ng hi\u1ec7n t\u1ea1i","operationId":"a7accc59384903af1b3597fc333f6d59","responses":{"200":{"description":"L\u1ea5y danh s\u00e1ch \u0111\u01a1n h\u00e0ng th\u00e0nh c\u00f4ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":true},"message":{"type":"string","example":"L\u1ea5y danh s\u00e1ch \u0111\u01a1n h\u00e0ng th\u00e0nh c\u00f4ng"},"data":{"type":"object"}},"type":"object"}}}},"401":{"description":"Ch\u01b0a x\u00e1c th\u1ef1c","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"Ch\u01b0a x\u00e1c th\u1ef1c"}},"type":"object"}}}},"500":{"description":"L\u1ed7i h\u1ec7 th\u1ed1ng","content":{"application\/json":{"schema":{"properties":{"status":{"type":"boolean","example":false},"message":{"type":"string","example":"L\u1ed7i h\u1ec7 th\u1ed1ng"}},"type":"object"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"securitySchemes":{"bearerAuth":{"type":"http","description":"Nh\u1eadp token JWT v\u00e0o \u0111\u00e2y","bearerFormat":"JWT","scheme":"bearer"},"0":{"bearerAuth":[]}}},"tags":[{"name":"X\u00e1c th\u1ef1c","description":"API cho \u0111\u0103ng nh\u1eadp v\u00e0 \u0111\u0103ng xu\u1ea5t"},{"name":"Qu\u1ea3n l\u00fd kh\u00e1ch h\u00e0ng","description":"API cho qu\u1ea3n l\u00fd kh\u00e1ch h\u00e0ng"}]}